Adding gmpy2 and PyCrypto to python.docker
Codewars does not have a ton of cryptography katas, but I hope to help change that because practical modern crypto is an extremely important topic. gmpy2 will help with high precision arithmetic needed when performing operations on very large numbers, such as in cryptography. However, it may also come in handy in other types of katas that involve working with extremely large numbers. PyCrypto will also give users access to basic crypto functions/algorithms that are missing in Python, since Python's built-in crypto is limited to the hashlib and hmac libraries, which is insufficient.